requirements.txt
absl-py==0.6.1
astor==0.7.1
backcall==0.1.0
cffi==1.11.5
colorama==0.4.1
cycler==0.10.0
decorator==4.3.0
dj-database-url==0.5.0
Django==2.1.3
django-bootstrap4==0.0.7
django-cors-headers==2.4.0
django-debug-toolbar==1.10.1
django-filter==2.0.0
django-heroku==0.3.1
django-jenkins==0.110.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
gast==0.2.0
grpcio==1.16.0
gunicorn==19.9.0
h5py==2.8.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.1
Keras==2.2.2
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
numpy==1.15.1
pandas==0.23.4
parso==0.3.1
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
protobuf==3.6.1
psycopg2==2.7.6.1
pycparser==2.19
Pygments==2.3.0
PyJWT==1.6.4
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
sqlparse==0.2.4
tensorboard==1.11.0
tensorflow==1.11.0
tensorflow-hub==0.1.1
tensorflowjs==0.6.5
termcolor==1.1.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
whitenoise==4.1
